fixes #5: Google OAuth signup session issue.
