--- src/App.css
@@ -1,3 +1,0 @@
-/* .ua-icon > svg:first-child {
-	display: none;
-} */
