+++ src/features/category/utils/format.ts
@@ -1,12 +1,16 @@
+import { Category } from '@/types';
+import { capitalize } from '@/utils/capitalize';
+
 export const humanize = (str: string): string => {
 	if (!str) return str;
 	const words = str.split('-');
-	return words
-		.map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-		.join(' ');
+	return words.map((word) => capitalize(word)).join(' ');
 };
 
-export const slugify = (str: string): string => {
-	if (!str) return str;
-	return str.toLowerCase().replace(/\s+/g, '-');
+export const slugify = (str: string): Category => {
+	return str
+		.toLowerCase()
+		.trim()
+		.replace(/[^a-z0-9]+/g, '-')
+		.replace(/^-+|-+$/g, '') as Category;
 };
