+++ src/types/auth.ts
@@ -88,19 +88,6 @@ export interface AuthActions {
 	logout: Logout;
 }
 
-export const categoryList = [
-	'cosmetics',
-	'electronics',
-	'apparels',
-	'home-appliances',
-	'furnitures',
-	'books',
-] as const;
-export type Category = (typeof categoryList)[number];
-export const isCategory = (str: string): str is Category => {
-	return categoryList.includes(str as Category);
-};
-
 export interface ErrorResponse {
 	status?: number;
 	statusText?: string;
